tg_support_site/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-telegram-bot==21.7
rsa==4.9
sniffio==1.3.1
starlette==0.41.2
Telethon==1.37.0
typing_extensions==4.12.2
uvicorn==0.32.0
websockets==14.0
